evaluation/llama_cpp/cpu/requirements.txt
accelerate
datasets >= 2.0
peft
protobuf
sentencepiece != 0.1.92
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
transformers
intel_extension_for_pytorch==2.2.0
git+https://github.com/huggingface/optimum.git@e38d40ad220a180213f99b1d93d0407a826c326d
optimum-intel
bitsandbytes  #baichuan
transformers_stream_generator
tiktoken  #qwen
einops  #qwen
git+https://github.com/intel/neural-speed.git@v1.0.1.dev0
auto-round
git+https://github.com/intel/neural-compressor.git
huggingface_hub
git+https://github.com/intel/intel-extension-for-transformers.git
lm-eval==0.4.2
llama-cpp-python[server]
